overview filter and auctions refactor
