+++ components/filters/FilterCard.vue
@@ -16,42 +16,6 @@
           </button>
         </div>
       </div>
-
-      <!--
-              <div class="space-y-2 mb-4">
-                <div v-if="filter.keywords.length > 0" class="flex flex-wrap gap-1">
-                  <span v-for="(keyword, i) in filter.keywords" :key="i"
-                    class="inline-flex items-center px-2 py-1 rounded-md text-xs font-medium bg-slate-100 dark:bg-slate-700 text-slate-700 dark:text-slate-300">
-                    {{ keyword }}
-                  </span>
-                </div>
-
-              <div v-if="filter.priceRange" class="flex items-center text-sm text-slate-500 dark:text-slate-400">
-                <TagIcon class="w-4 h-4 mr-1" />
-                <span>{{ filter.priceRange }}</span>
-              </div>
-
-
-              <div v-if="filter.location" class="flex items-center text-sm text-slate-500 dark:text-slate-400">
-                <MapPinIcon class="w-4 h-4 mr-1" />
-                <span>{{ filter.location }}</span>
-              </div>
-            </div>
-
-            <div class="flex items-center justify-between pt-3 border-t border-slate-200 dark:border-slate-700">
-              <div class="flex items-center space-x-2">
-                <span class="inline-flex items-center px-2 py-1 rounded-full text-xs font-medium"
-                  :class="filter.active ? 'bg-green-100 text-green-800 dark:bg-green-900 dark:text-green-200' : 'bg-slate-100 text-slate-800 dark:bg-slate-700 dark:text-slate-200'">
-                  {{ filter.active ? 'Active' : 'Paused' }}
-                </span>
-                <span class="text-sm text-slate-500 dark:text-slate-400">{{ filter.matchCount }} matches</span>
-              </div>
-              <NuxtLink :to="`/auctions?filter=${filter.id}`"
-                class="text-sm text-indigo-600 dark:text-indigo-400 hover:underline">
-                View matches
-              </NuxtLink>
-            </div>
-            -->
     </div>
   </div>
 </template>
