+++ locales/en.json
@@ -69,6 +69,7 @@
   "landingHero1": "Never Miss a",
   "landingHero2": "Deal",
   "landingHero3": "Again",
+  "loadmo": "Loading more auctions...",
   "manFilt": "Manage your auction search filters",
   "market": "Marketplace",
   "matchAuc": "Matched Auctions",
