+++ pages/auctions/AuctionsHeader.vue
@@ -1,0 +1,8 @@
+<template>
+  <div class="flex flex-col md:flex-row justify-between items-start md:items-center mb-8">
+    <div>
+      <UiHeaderLabel :label="$t('matchAuc')" />
+      <p class="mt-2 text-slate-500 dark:text-slate-400">{{ $t('aucFil') }}</p>
+    </div>
+  </div>
+</template>
