+++ types/globals.d.ts
@@ -4,3 +4,15 @@ interface topFilter {
   keywords: string[];
 }
 
+interface filterFace {
+  id: number;
+  name: string;
+  marketplace: string;
+  keywords: string[];
+  priceRange: string;
+  location: string;
+  active: boolean;
+  matchCount: number;
+}
+
+
